pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "change-analysis"
version = "1.0.0"
description = "Production-ready module for web page change detection and importance analysis"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Platform Engineering Team"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "Pillow>=10.0.0",
    "lxml>=4.9.0",
    "beautifulsoup4>=4.11.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
aws = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "boto3>=1.28.0",
    "botocore>=1.31.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/change-analysis"
Documentation = "https://github.com/yourorg/change-analysis/blob/main/README.md"
Repository = "https://github.com/yourorg/change-analysis"

[tool.setuptools]
packages = ["change_analysis", "api"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
change_analysis = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
